some bugs fixed

+++ src/controller/blogController.js
@@ -7,7 +7,7 @@ const isValidObjectId = (ObjectId) => {
 const stringChecking = function (data) {
     if (typeof data !== 'string') {
         return false;
-    } else if (typeof data === 'string' && data.trim().length == 0) {
+    } else if (typeof data === 'string' && data.trim().length === 0) {
         return false;
     } else {
         return true;
@@ -31,12 +31,12 @@ const createBlogs = async function (req, res) {
     try {
         let bodyData = req.body
         if (Object.keys(bodyData).length === 0) {
-            return res.status(400).send({ status: false, msg: "Please enter details for creation of blog" })
+            return res.status(400).send({ status: false, msg: "Please enter details for creating a blog" })
         }
         else {
             const { title, authorId, body, subcategory, category, tags } = bodyData
 
-            if (!isValidObjectId(authorId)) return res.status(400).send({ status: false, msg: "please enter valid authorId" })
+            if (!isValidObjectId(authorId)) return res.status(400).send({ status: false, msg: `Author Id: ${authorId} is not valid, please enter valid authorId` })
 
             if (!stringChecking(title)) return res.status(400).send({ status: false, msg: "title  must be present and have Non empty string " })
 
@@ -77,29 +77,26 @@ const getBlog = async function (req, res) {
         let bodyData = req.query
 
         if (Object.keys(bodyData).length == 0) {
-            let getData = await blogModel.find({ isDeleted: false, isPublished: true })
+            let getData = await blogModel.find({ isDeleted: false, deletedAt: null, isPublished: true })
             if (getData.length <= 0) {
-                return res.status(404).send({ status: false, msg: "Data Not Found" })
+                return res.status(404).send({ status: false, msg: "No entries are there to show you!" })
             }
             return res.status(200).send({ status: true, count: getData.length, data: getData })
         }
         else {
-            let { subcategory, category, tags, authorId } = bodyData
-            let filter = {}
+            const { subcategory, category, tags, authorId } = bodyData
+            const filter = {}
             if (subcategory) {
-                if (!arrayOfStringChecking(subcategory)) return res.status(400).send({ status: false, msg: "subcategory must be present and have Non empty string " })
                 filter.subcategory = subcategory
             }
             if (tags) {
-                if (!arrayOfStringChecking(tags)) return res.status(400).send({ status: false, msg: "tags must be present and have Non empty string " })
                 filter.tags = tags
             }
             if (category) {
-                if (!arrayOfStringChecking(category)) return res.status(400).send({ status: false, msg: "tags must be present and have Non empty string " })
                 filter.category = category
             }
             if (authorId) {
-                if (!isValidObjectId(authorId)) return res.status(400).send({ status: false, msg: "please enter valid authorId" })
+                if (!isValidObjectId(authorId)) return res.status(400).send({ status: false, msg: `AuthorId: ${authorId} is not valid, please enter valid authorId` })
                 filter.authorId = authorId
             }
             filter.isDeleted = false
@@ -122,7 +119,7 @@ const updateBlogs = async function (req, res) {
     try {
         let blogId = req.params.blogId
         let data = req.body
-        let { subcategory, tags, body, title } = data
+        const { subcategory, tags, body, title } = data
         if (Object.keys(data).length === 0) {
             return res.status(400).send({ status: false, msg: "Please enter required details in request body" })
         }
@@ -152,7 +149,7 @@ const updateBlogs = async function (req, res) {
                 { "$set": { "title": title, "body": body }, "$addToSet": { "tags": tags, "subcategory": subcategory }, isPublished: true, publishedAt: new Date() },
                 { new: true }
             )
-            return res.status(200).send({ status: true, data: updateUser })
+            return res.status(200).send({ status: true, message: "Updated Successfully", data: updateUser })
         }
 
     } catch (error) {
@@ -173,7 +170,7 @@ const deleteBlog = async function (req, res) {
             await blogModel.findByIdAndUpdate(blogId, {
                 $set: { isDeleted: true, deletedAt: new Date() },
             });
-            return res.status(200).send({ status: false, msg: "Deleted" });
+            return res.status(200).send({ status: true, msg: "Deleted" });
         }
     }
     catch (err) {
@@ -193,15 +190,12 @@ const deleteByQuery = async function (req, res) {
         const { category, subcategory, tags, authorId } = data
         const filter = {}
         if (subcategory) {
-            if (!arrayOfStringChecking(subcategory)) return res.status(400).send({ status: false, msg: "subcategory must be present and have Non empty string " })
             filter.subcategory = subcategory
         }
         if (tags) {
-            if (!arrayOfStringChecking(tags)) return res.status(400).send({ status: false, msg: "tags must be present and have Non empty string " })
             filter.tags = tags
         }
         if (category) {
-            if (!arrayOfStringChecking(category)) return res.status(400).send({ status: false, msg: "tags must be present and have Non empty string " })
             filter.category = category
         }
         if (authorId) {
@@ -215,32 +209,29 @@ const deleteByQuery = async function (req, res) {
         filter.isPublished = false
         if (subcategory || category || tags || authorId) {
             const blog = await blogModel.find(filter)
+            console.log(blog)
+            if(blog.length === 0) return res.status(403).send({ status: false, msg: "Blog not found" })
             for (let i = 0; i < blog.length; i++) {
-                if (((blog[i].authorId.toString())) !== tokenAuthorId) {
-                    return res.status(403).send({ status: false, msg: "The blogs with this filters are forbidden for this logged in user" })
+                let blogAuthorId = blog[i].authorId.toString()
+                if (blogAuthorId === tokenAuthorId) {
+                    filter.authorId = blogAuthorId
+                    const deleteByQuery = await blogModel.updateMany(
+                        filter,
+                        { $set: { isDeleted: true, deletedAt: new Date() } }
+                    )
+                    if (deleteByQuery.modifiedCount === 0) {
+                        return res.status(404).send({ status: false, msg: "Nothing to delete, already deleted" })
+                    }
+                    return res.status(200).send({ status: true, msg: "deleted" })
                 }
             }
-            const deleteByQuery = await blogModel.updateMany(
-                filter,
-                { $set: { isDeleted: true, deletedAt: new Date() } }
-            )
-            if (deleteByQuery.modifiedCount === 0) {
-                return res.status(404).send({ status: false, msg: "No data found" })
-            }
-            return res.status(200).send({ status: true, msg: "deleted" })
         } else {
             return res.status(400).send({ status: false, msg: "filters can be subcategory, category, tags, authorId only " })
         }
-
-
     } catch (error) {
         return res.status(500).send({ status: false, error: error.message });
     }
 }
 
 
-module.exports.createBlogs = createBlogs
-module.exports.getBlog = getBlog
-module.exports.updateBlogs = updateBlogs
-module.exports.deleteByQuery = deleteByQuery
-module.exports.deleteBlog = deleteBlog
+module.exports = { createBlogs, getBlog, updateBlogs, deleteBlog, deleteByQuery }
